[feat/#43] homeactivity 병함 (guiactivity 삭제)
